# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_MemoryItemResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MemoryItemResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "default": "success"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[MemoryItemResponse]",
        "type": "object"
      },
...
      "ApiResponse_list_MemoryItemResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MemoryItemResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "default": "success"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[MemoryItemResponse]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryCreateRequest": {
        "description": "Client-supplied fields for creating a Memory Item.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "referenced_entities": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Referenced Entities",
            "type": "array"
          },
          "referenced_findings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Referenced Findings",
            "type": "array"
          },
          "referenced_relationships": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Referenced Relationships",
            "type": "array"
          },
          "source_investigation_id": {
            "title": "Source Investigation Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MemoryStatus"
          },
          "supporting_evidence": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Supporting Evidence",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "source_investigation_id",
          "confidence",
          "status"
        ],
        "title": "MemoryCreateRequest",
        "type": "object"
      },
      "MemoryItemResponse": {
        "description": "Flat projection of a Memory Item.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "referenced_entities": {
            "items": {
              "type": "string"
            },
            "title": "Referenced Entities",
            "type": "array"
          },
          "referenced_findings": {
            "items": {
              "type": "string"
            },
            "title": "Referenced Findings",
            "type": "array"
          },
          "referenced_relationships": {
            "items": {
              "type": "string"
            },
            "title": "Referenced Relationships",
            "type": "array"
          },
          "source_investigation_id": {
            "title": "Source Investigation Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "supporting_evidence": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Evidence",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "type",
          "source_investigation_id",
          "confidence",
          "status",
          "created_at",
          "version",
          "content",
          "supporting_evidence",
          "referenced_findings",
          "referenced_entities",
          "referenced_relationships"
        ],
        "title": "MemoryItemResponse",
        "type": "object"
      },
      "MemoryStatus": {
        "description": "Lifecycle states of a Memory Item.\n\n``DEPRECATED`` controls *relevance* (retrieval-invisible knowledge);\n``ERASED`` is the terminal *end-of-life* state (data-lifecycle.md \u00a73\n\"deprecation is not deletion\", ADR-017): its content is redacted and its\nderived embedding is deleted through the outbox projection.",
        "enum": [
          "candidate",
          "verified",
          "organizational",
          "deprecated",
          "erased"
        ],
        "title": "MemoryStatus",
        "type": "string"
      },
      "MemoryUpdateRequest": {
        "description": "Full Memory Item representation supplied on update (PUT semantics).\n\nThe identifier is never part of the body: it is taken from the path. The\n``version`` must immediately follow the latest persisted version.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "referenced_entities": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Referenced Entities",
            "type": "array"
          },
          "referenced_findings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Referenced Findings",
            "type": "array"
          },
          "referenced_relationships": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Referenced Relationships",
            "type": "array"
          },
          "source_investigation_id": {
            "title": "Source Investigation Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MemoryStatus"
          },
          "supporting_evidence": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Supporting Evidence",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "type",
          "source_investigation_id",
          "confidence",
          "status",
          "version",
          "created_at"
        ],
        "title": "MemoryUpdateRequest",
        "type": "object"
      },
...
      "ResponseMeta": {
        "description": "Response metadata common to success and error responses.",
        "properties": {
          "api_version": {
            "title": "Api Version",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "correlation_id",
          "timestamp",
          "api_version"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
      "ResponseStatus": {
        "description": "The outcome status carried by every API response.",
        "enum": [
          "success",
          "error"
        ],
        "title": "ResponseStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SentinelAI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/memory": {
      "get": {
        "description": "List the latest Memory Item versions originating from an investigation.\n\nThe identifier is a structural cross-service reference\n(database-architecture \u00a78a): an unknown investigation yields an empty\nlist, not an error. Memory is the shared knowledge layer (\u00a76a), so the\nlisting is open to any authenticated identity.",
        "operationId": "list_memory_api_v1_memory_get",
        "parameters": [
          {
            "in": "query",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_MemoryItemResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Memory",
        "tags": [
          "memory"
        ]
      },
      "post": {
        "operationId": "create_memory_api_v1_memory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MemoryItemResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/{memory_id}": {
      "delete": {
        "description": "Erase a Memory Item \u2014 the person-linked right-to-be-forgotten path.\n\nRequires the ``knowledge:erase`` capability (ADR-019): the promotion\nboundary explains who may *read* organizational knowledge and says nothing\nabout who may end its existence, so destroying it is gated rather than\nopen to any authenticated identity. The authorization seam enforces it\nbefore this endpoint runs.\n\nDistinct from deprecation, which controls relevance rather than existence:\nevery version is tombstoned and the derived embedding is deleted through\nthe outbox projection. Idempotent \u2014 re-erasing returns the tombstone.",
        "operationId": "erase_memory_api_v1_memory__memory_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MemoryItemResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Erase Memory",
        "tags": [
          "memory"
        ]
      },
      "get": {
        "operationId": "get_memory_api_v1_memory__memory_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MemoryItemResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Memory",
        "tags": [
          "memory"
        ]
      },
      "put": {
        "operationId": "update_memory_api_v1_memory__memory_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MemoryItemResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/{memory_id}/deprecate": {
      "post": {
        "operationId": "deprecate_memory_api_v1_memory__memory_id__deprecate_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MemoryItemResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deprecate Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/{memory_id}/history": {
      "get": {
        "operationId": "get_memory_history_api_v1_memory__memory_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_MemoryItemResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Memory History",
        "tags": [
          "memory"
        ]
      }
    },
...
  }
}